requirements.txt
beautifulsoup4==4.9.3
openpyxl==3.0.7
requests==2.25.1
rpaframework==10.6.0